backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pymysql==1.1.1
cryptography==43.0.1
celery==5.4.0
redis==5.1.1
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic-settings==2.5.2
httpx==0.27.2
pytesseract==0.3.13
Pillow==10.4.0
aiofiles==24.1.0
mt-940>=4.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
